Remove circular dependency.
