+++ server/utils/index.js
@@ -1,5 +1,4 @@
 export * from '../../universal/utils';
-export * from './server';
 export * from './flash';
 export * from './hardware';
 export * from './logging';
